+++ js/game.js
@@ -581,6 +581,8 @@ var Game = function(opt)
 
         if(!looping)
         {
+            audio.setAttribute('volume_target_type', target);
+            target = game.gui.get_value(target) * game.gui.get_value('global_volume');
             audio.volume=0;
         }
         if(audio.volume<target)
@@ -604,6 +606,7 @@ var Game = function(opt)
         else
         {
             audio.pause()
+            audio.removeAttribute('volume_target_type');
         }
     };